oonipipeline/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oonipipeline"
dynamic = ["version"]
description = ""
authors = [{ name = "OONI", email = "contact@ooni.org" }]

dependencies = [
  "boto3 ~= 1.24",
  "PyYAML ~= 6.0",
  "tqdm ~= 4.64",
  "lz4 ~= 4.0",
  "requests ~= 2.27",
  "cryptography ~= 41.0.0",
  "clickhouse-driver ~= 0.2",
  "click ~= 8.0.0",
  "lxml ~= 4.9",
  "maxminddb ~= 2.2",
  "orjson ~= 3.8",
  "mashumaro ~= 3.0",
  "fastapi ~= 0.108.0",
  "tabulate ~= 0.9.0",
  "warcio ~= 1.7.4",
  "msgpack ~= 1.0.4",
  "click-loglevel ~= 0.5.0",
  "temporalio ~= 1.7.0",
  "temporalio[opentelemetry] ~= 1.7.0",
  "opentelemetry-exporter-otlp-proto-grpc ~= 1.29.0",
  "uvicorn ~= 0.25.0",
  "pydantic-settings ~= 2.4.0",
  "apache-airflow == 2.10.4"
]

[project.optional-dependencies]
analysis = [
  "altair ~= 4.2.0",
  "pandas ~= 2.0.0",
  "jupyterlab ~= 4.0.7",
  "bokeh ~= 3.5.2",
]

[tool.hatch.build.targets.sdist]

[tool.hatch.build.targets.wheel]
packages = ["src/oonipipeline", "src/oonidata"]


[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
  "black",
  "snakeviz",
  "mypy",
  "pytest-benchmark",
  "pytest-profiling",
  "memray",
  "viztracer",
  "pytest-docker",
  "pytest-asyncio",
  "ipdb",
]
python = "3.11"
path = ".venv/"

[tool.hatch.version]
path = "src/oonipipeline/__about__.py"

[tool.hatch.envs.default.scripts]
oonipipeline = "python -m oonipipeline.main {args}"
dataviz = "uvicorn oonipipeline.dataviz.main:app {args}"
api = "uvicorn oonipipeline.api.main:app {args}"
test = "pytest {args:tests}"
# --full-trace --log-level=INFO  --log-cli-level=INFO -v --setup-show -s
test-cov = "pytest --cov=./ --cov-report=xml --cov-report=html --cov-report=term {args:tests}"
cov-report = ["coverage report"]
cov = ["test-cov", "cov-report"]
